Time: 19 ms (12.79%), Space: 35 MB (34.08%) - LeetHub

+++ 1011-capacity-to-ship-packages-within-d-days/1011-capacity-to-ship-packages-within-d-days.cpp
@@ -9,7 +9,7 @@ public:
             int count = 0;
             int i = 0;
             while (i < n) {
-                long long cost = 0;
+                int cost = 0;
                 while (i < n && cost + nums[i] <= mid) {
                     cost += nums[i];
                     i++;
